+++ config/lighthouse-config/package.json
@@ -21,7 +21,7 @@
     "clean": "rm -rf .turbo && rm -rf dist",
     "clean:install": "pnpm run clean && rm -rf node_modules",
     "copy": "mkdir -p ./dist && ./scripts/copy.sh",
-    "dev": "echo '# tsup-node  --watch'",
+    "dev": "echo '# tsdown  --watch'",
     "format:lint": "pnpm run lint:eslint --fix && pnpm run format:prettier --write",
     "format:prettier": "prettier \"./src/**/*.{cjs,js,jsx,mjs,ts,tsx,json,md,mdx,css,html,yml,yaml,scss}\" --ignore-unknown --log-level warn",
     "lint": "pnpm run format:prettier --check && pnpm run lint:eslint",
