+++ config/lint-staged/tsdown.config.ts
@@ -1,0 +1,17 @@
+import type { Options } from 'tsdown'
+
+import { defineConfig } from 'tsdown'
+
+import { config as _config } from '../../tsdown.config.ts'
+
+const entry = ['src/*.ts']
+
+const config: Options = {
+  ..._config,
+  dts: false,
+  entry,
+}
+
+export default defineConfig({
+  ...config,
+})
