+++ package.json
@@ -56,7 +56,7 @@
     "lodash": "4.17.21",
     "semantic-release": "24.2.7",
     "syncpack": "13.0.4",
-    "tsup": "8.5.0",
+    "tsdown": "0.14.2",
     "turbo": "2.5.6",
     "typescript": "5.9.2"
   },
