+++ packages/design-system/tsdown.config.ts
@@ -1,9 +1,8 @@
-// import isCI from 'is-ci'
-import type { Options } from 'tsup'
+import type { Options } from 'tsdown'
 
-import { defineConfig } from 'tsup'
+import { defineConfig } from 'tsdown'
 
-import { config as _config } from '../../tsup.config.js'
+import { config as _config } from '../../tsdown.config.ts'
 import pkg from './package.json'
 
 /**
@@ -40,7 +39,7 @@ const entry = [
 ]
 
 /**
- * @todo(tsup) tsup-node does this automatically we can remove
+ * @todo(tsup) tsdown does this automatically we can remove
  */
 const external = [
   ...Object.keys(pkg.dependencies || {}),
@@ -56,7 +55,6 @@ const config: Options = {
   external,
   inject,
   minify: true,
-  splitting: true,
   tsconfig: 'tsconfig.json',
 }
