--- packages/release-notes-generator/tsup.config.ts
@@ -1,15 +1,0 @@
-import type { Options } from 'tsup'
-
-import { defineConfig } from 'tsup'
-
-import { config as _config } from '../../tsup.config.js'
-
-const entry = ['src/*']
-const config: Options = {
-  ..._config,
-  entry,
-}
-
-export default defineConfig({
-  ...config,
-})
