+++ packages/spotify/tsdown.config.ts
@@ -1,9 +1,9 @@
 // import isCI from 'is-ci'
-import type { Options } from 'tsup'
+import type { Options } from 'tsdown'
 
-import { defineConfig } from 'tsup'
+import { defineConfig } from 'tsdown'
 
-import { config as _config } from '../../tsup.config.js'
+import { config as _config } from '../../tsdown.config.ts'
 
 const entry = ['src/**']
 const config: Options = {
