+++ tsdown.config.ts
@@ -1,23 +1,16 @@
-/**
- * @ref(tsup) https://tsup.egoist.dev/
- *
- */
-import type { Options } from 'tsup'
+import type { Options } from 'tsdown'
 
-import { defineConfig } from 'tsup'
+import { defineConfig } from 'tsdown'
 
 const config: Options = {
-  clean: true,
   dts: true,
   external: [],
-  format: ['esm'],
   minify: true,
   noExternal: [],
   onSuccess: 'pnpm run copy',
   outDir: 'dist',
   silent: true,
   sourcemap: false,
-  splitting: false,
   target: ['node22'],
   treeshake: false,
 }
